EX-2465 Handle `null` in text functions; remove bessel
In Excel, functions that expect a value but instead receive no value
give the `N/A` error. The text functions in package.js were not
checking for `null` values, so all sorts of errors were resulting. This
commit adds checks for `null` to all functions that take a text input
and did not have that check.

Adds a utility function `isAnyNull`.

Bessel functions are not officially supported (they are not listed in
the Zendesk documents), so those were removed.

Minor linter cleanup.